Consolidate shared color and distance helpers
Extract toRgbaColor into src/utils/cssColor.ts so the duplicate
copies in maplibreRangeRings.ts and maplibreScenarioFeatures.ts can
share one implementation. Move distanceMeters and isZeroLengthSegment
out of maplibreTouchDrawing.ts into a new src/geo/distance.ts module
and update consumers. Tighten the OL stub typing in
ScenarioEditorMaplibre.vue and drop the redundant double parseCssColor
fallback.
