+++ src/composables/maplibreTouchDrawing.ts
@@ -1,6 +1,6 @@
 import type { Position } from "geojson";
+import { isZeroLengthSegment } from "@/geo/distance";
 
-const ZERO_LENGTH_SEGMENT_TOLERANCE_METERS = 0.01;
 const TOUCH_DOUBLE_TAP_MS = 350;
 const TOUCH_DOUBLE_TAP_TOLERANCE_PX = 30;
 
@@ -63,23 +63,6 @@ export function getPreviousDistinctVertex(coordinates: Position[], beforeIndex: 
   return null;
 }
 
-export function isZeroLengthSegment(a: Position, b: Position) {
-  return distanceMeters(a, b) <= ZERO_LENGTH_SEGMENT_TOLERANCE_METERS;
-}
-
-export function distanceMeters(a: Position, b: Position) {
-  if (a[0] === b[0] && a[1] === b[1]) return 0;
-  const earthRadius = 6_371_008.8;
-  const phi1 = degreesToRadians(a[1]);
-  const phi2 = degreesToRadians(b[1]);
-  const deltaPhi = degreesToRadians(b[1] - a[1]);
-  const deltaLambda = degreesToRadians(b[0] - a[0]);
-  const h =
-    Math.sin(deltaPhi / 2) ** 2 +
-    Math.cos(phi1) * Math.cos(phi2) * Math.sin(deltaLambda / 2) ** 2;
-  return 2 * earthRadius * Math.atan2(Math.sqrt(h), Math.sqrt(1 - h));
-}
-
 export function suppressMapEvent(e: SuppressibleMapEvent) {
   e.preventDefault();
   if (canPreventDomDefault(e.originalEvent)) {
@@ -102,8 +85,4 @@ function getEventTime(e: TouchEventLike) {
 
 function distancePx(a: { x: number; y: number }, b: { x: number; y: number }) {
   return Math.hypot(a.x - b.x, a.y - b.y);
-}
-
-function degreesToRadians(value: number) {
-  return (value * Math.PI) / 180;
 }
